inpaint/requirements.txt
absl-py==2.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
albumentations==0.5.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
braceexpand==0.1.7
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
-e git+https://github.com/DepthAnything/Depth-Anything-V2.git@e5a2732d3ea2cddc081d7bfd708fc0bf09f812f1#egg=depth_anything_v2
easydict==1.9
exceptiongroup==1.3.0
fastapi==0.115.13
ffmpy==0.6.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
gradio==4.29.0
gradio_client==0.16.1
gradio_imageslider==0.0.20
grpcio==1.73.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imgaug==0.4.0
importlib_resources==6.5.2
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.1
jpeg4py==0.1.4
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.8
kornia==0.5.0
lazy_loader==0.4
libclang==18.1.1
lmdb==1.6.2
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.5.0
namex==0.1.0
narwhals==1.43.1
networkx==3.4.2
numpy==1.24.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pillow==10.4.0
portalocker==3.2.0
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==1.2.9
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
ruff==0.12.0
safetensors==0.5.3
-e git+https://github.com/facebookresearch/sam2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4#egg=SAM_2
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
-e git+https://github.com/geekyutao/Inpaint-Anything.git@5bfa9f3d1a829d71bddecf28c228952dbf825e3a#egg=segment_anything&subdirectory=segment_anything
semantic-version==2.10.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.15
tomlkit==0.12.0
torch==2.7.1
torchaudio==2.7.1
torchmetrics==0.2.0
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
webdataset==1.0.2
websockets==11.0.3
Werkzeug==3.1.3
wldhx.yadisk-direct==0.0.6
wrapt==1.17.2
yarl==1.20.1
